create a new view system (WIP)
